Created initial structure for the graphics subsystem
- Added JUnit 4
- Added test folder to classpath
- Created empty shells for Entity and Tank, since these are needed for
sprites. The implementation and interfaces are left for the data team.
- Added Sprite interface, and simple TankSprite class. More to come once
we select our graphics library...
- Added DrawLayer enum, which is used to specify the draw order
- Added tests for TankSprite and Tank. Entity is abstract and has no
methods, so there is nothing to test. Sprite<T> is an interface.
DrawLayer is an enum without methods, so again, there is nothing to
test.
- Note that there are some warnings for the tests. This should be fine,
since tests sometimes need to do things (like unused allocations) that
would be incorrect in production code. This does not impact the quality
of our application, since tests aren't production code.