Added comments to new UI classes
If we weren't scrambling to get this done before Sprint 1 ends, I'd have
liked to spend more time cleaning up this code.

As it stands, we need this basic functionality to meet our minimum
requirements. We should revisit this code after our rush to finish
Sprint 1 ends and make the necessary changes to bring it up to our
standards.