Implemented Review Feedback on #274
* Commented out getOverlappingEntities, for now, since it isn't used,
but the basic idea is there.
* Added a self-check to Tank.getLookaheadEntities and set Tank's to be
solid in their constructor.
* Effect.update should have a unit test already -- it's in the
EffectTest class, in test.bubolo.world.entity.
* AdaptiveTileUtil now has a description.
* adaptiveSplit_9 removed from TextureUtil, since it was used for
DeepWater only and that has been replaced with a better, prettier
method.
* Entity.setSolid visibility changed to protected.
* Changed ColorSets ints to finals. I'm not super familiar with how
enums work, so if someone wants to switch this over to enums in the
future (or tell me how), that'd be great!